+++ packages/moonstone/VideoPlayer/VideoPlayer.js
@@ -284,14 +284,6 @@ const VideoPlayerBase = class extends React.Component {
 		loading: PropTypes.bool,
 
 		/**
-		 * An object containing properties to be passed to media controls component.
-		 *
-		 * @type {Object}
-		 * @public
-		 */
-		mediaControlsProps: PropTypes.Object,
-
-		/**
 		 * Amount of time (in milliseconds), after the last user action, that the `miniFeedback`
 		 * will automatically hide.
 		 * Setting this to 0 or `null` disables `miniFeedbackHideDelay`; `miniFeedback` will always
@@ -1908,7 +1900,6 @@ const VideoPlayerBase = class extends React.Component {
 			jumpForwardIcon,
 			leftComponents,
 			loading,
-			mediaControlsProps,
 			moreButtonCloseLabel,
 			moreButtonDisabled,
 			moreButtonLabel,
@@ -2055,7 +2046,6 @@ const VideoPlayerBase = class extends React.Component {
 							</MediaSlider>}
 
 							<MediaControls
-								{...mediaControlsProps}
 								backwardIcon={backwardIcon}
 								forwardIcon={forwardIcon}
 								jumpBackwardIcon={jumpBackwardIcon}